Loop through members until creating users is succesful fix #2
While users are not created only localhost connections can be used
because all other connections require authentication. Hence, we can
not simply let the mongo CLI find the master and then execute the
add user command.

Therefor if member-0 is not the master we loop through member pods
until creating the users was succesful.